chore: assign real IDs to new iOS components
