+++ demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0114/script.js
