+++ demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0115/script.js
