+++ tests/ios/MASVS-RESILIENCE/MASTG-TEST-0091.md
@@ -9,7 +9,7 @@ masvs_v1_levels:
 - R
 profiles: [R]
 status: deprecated
-covered_by: [MASTG-TEST-0x01]
+covered_by: [MASTG-TEST-0350]
 deprecation_note: "New version available in MASTG V2"
 ---
